requirements.txt
blinker==1.6.2
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-loading-spinners==1.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
duckdb==0.7.1
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
Werkzeug==2.3.3
